pyproject.toml
[tool.poetry]
name = "tcc-mba"
version = "0.1.0"
description = ""
authors = ["Bruno Kim <brunokim.mc@gmail.com>"]
readme = "README.md"
packages = [{include = "tcc_mba"}]

[tool.poetry.dependencies]
attrs = "^23.1.0"
cattrs = "^23.1.2"
db-dtypes = "^1.2.0"
google-cloud-bigquery = "^3.20.1"
google-cloud-storage = "^2.16.0"
itables = "^2.0.0"
jupyterlab = "^4.1.6"
numpy = "^1.26.4"
pandas = "^2.2.1"
pyarrow = "^15.0.2"
python = "^3.10"
ruamel-yaml = "^0.17.32"
seaborn = "^0.13.2"
statsmodels = "^0.14.1"
tqdm = "^4.66.1"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
